pyproject.toml
[project]
name = "APOEZ_Website_Backend"
version = "1.0.0"
description = "backend to handle lost and found, backtests, and loaner teach"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiocache>=0.12.3",
    "fastapi-sso>=0.17.0",
    "fastapi[standard]>=0.115.6",
    "bleach>=6.1.0",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "pymongo>=4.13.2",
    "ruff>=0.9.2",
    "sentry-sdk[fastapi]>=2.20.0",
    "valkey-glide>=2.0.1",
    "pytest>=8.3.3",
]

[tool.ruff]
line-length = 88
fix = true

[tool.ruff.lint]
select = ["I", "F", "E"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

